// js/stat-evaluation-player/src/generated/MustyFlickEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type MustyFlickEvent = { time: number, frame: number, player: RemoteIdTs, is_team_0: boolean, aerial: boolean, dodge_time: number, dodge_frame: number, time_since_dodge: number, confidence: number, local_ball_position: [number, number, number], rear_alignment: number, top_alignment: number, forward_approach_speed: number, pitch_rate: number, ball_speed_change: number, };