// js/stat-evaluation-player/src/generated/MustyFlickStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledCounts } from "./LabeledCounts.ts";

export type MustyFlickStats = { count: number, aerial_count: number, high_confidence_count: number, is_last_musty: boolean, last_musty_time: number | null, last_musty_frame: number | null, time_since_last_musty: number | null, frames_since_last_musty: number | null, last_confidence: number | null, best_confidence: number, cumulative_confidence: number, labeled_event_counts?: LabeledCounts, };